bindings: Rework the parsing logic
Now we just create/get the binding when we know the type, instead of
relying of invalid values to detect the case later. The only downside is
that we may create a scope with no bindings when parsing may fail.

This way, we can (later) use the button == 0 value as a correct value.

Signed-off-by: Quentin Glidic <sardemff7+git@sardemff7.net>