token: Use no SI prefix for zero value
When zero value was passed to _nk_token_list_append_prettify(), the
search for a proper SI prefix in _nk_token_prefixes_si_small run out
of array bounds. In my case, it manifested as bitrate string of
"0(null)b/s".

This is now fixed by explicit check for zero value and setting of
empty prefix in this case.

Signed-off-by: Michal Sojka <michal.sojka@cvut.cz>